Refactor existing code
This changes List and Row to be far dumber. Row now only knows a number
and an item and renders them out to the screen - it doesn't have any
knowledge of fetching data.

List has also been changed. While it still has a fair amount of logic
a lot of it has been moved out of the render method and given clearer
names to make it easy to see what's going on. List still only deals with
calculating which rows need fetching - the ability to do the fetching is
still stored outside the component, but that's now provided by a parent
component and passed down through the props.