Add caching wrapper
This adds a new Cache component, inserted between the App and the List.
It overrides the `loadData` method passed to list to add caching
capabilities.

If some of the required data isn't available in the cache it loads all
of it. If the external server is significantly slower to return larger
amounts of data it could be worth determining which data is missing and
trying to fetch smaller batches to fill in the holes in the cache.

I believe that in most cases the overhead of making multiple network
requests if there are several non-concurrent sets of uncached data would
make any performance gained from requesting a smaller amount of data
negligible.