Switch the data format to an object
Using an object, with the row number as the key and the name as the
value simplifies quite a bit. The top level App component is the only
place we have to deal with converting the array we get from the backend
to an object and this will enable us to batch requests in future.