Batch requests together
This adds another component that wraps the Cache to batch requests
together to prevent regular scrolling from hitting the backend with too
many requests. The Batcher takes the amount of time to wait for
subsequent requests in as a prop. This timeout is reset whenever new
data is added to the batch. When the timer expires it makes a single
request for all the data that's been requested in this batch.